requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
compress-pickle==2.1.0
contourpy==1.0.7
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
dominate==2.7.0
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
html5lib==1.1
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.4.1
importlib-resources==5.12.0
imutils==0.5.4
iniconfig==2.0.0
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
Keras-Applications==1.0.8
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lit==16.0.1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pdfkit==1.0.0
Pillow==9.5.0
pluggy==1.0.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.7
pydantic-numpy==1.4.4
pydyf==0.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyphen==0.14.0
pytesseract==0.3.10
pytest==7.3.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scipy==1.9.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
weasyprint==58.1
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
zopfli==0.2.2
